+++ src/test/_setup.ts
@@ -2,10 +2,9 @@
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
 import * as chai from 'chai';
-
-import { SharedPool } from '../shared-pool';
+import { ConnectionPool } from '../connection-pool';
 
 chai.use(require('chai-subset')); // tslint:disable-line
 chai.use(require('chai-as-promised')); // tslint:disable-line
 
-(SharedPool as any).deterministicInsertion = true;
+ConnectionPool.deterministicOrder = true;